Expand SurfScoreboard functionalities and documentation
This commit significantly extends the functionalities of the SurfScoreboard in the SurfBukkitPacketApi. It includes the ability to handle varying maximum lines, dynamic content, animations and automatic updates. It also introduces a simpler viewer management by making scoreboards globally accessible. Furthermore, detailed documentation for all methods is provided for better understanding and ease of use.